--- src/main/java/racingCar/Entity/RacingCar.java
@@ -1,17 +1,0 @@
-package racingCar.Entity;
-
-public class RacingCar {
-    private int position;
-
-    public RacingCar(int position) {
-        this.position = position;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-
-    public void setPosition(int position) {
-        this.position = position;
-    }
-}
