+++ src/main/java/racingCar/domain/Movable.java
@@ -1,0 +1,5 @@
+package racingCar.domain;
+
+public interface Movable {
+    public void move(int randomNumber);
+}
